Move extract func and its tests to clients dux module
