+++ src/helpers/tests/utils.test.tsx
@@ -1,7 +1,5 @@
 import * as gatekeeper from '@onaio/gatekeeper';
 import { ONADATA_OAUTH_STATE, OPENSRP_OAUTH_STATE } from '../../configs/env';
-import * as rawClientFixture from '../../services/clients/tests/fixtures';
-import { extractClient } from '../utils';
 
 import { oAuthUserInfoGetter } from '../utils';
 
@@ -29,20 +27,5 @@ describe('helpers/utils', () => {
     oAuthUserInfoGetter(resp);
     expect(mock).toHaveBeenCalledWith(resp);
     mock.mockRestore();
-  });
-
-  it('extracts clients correctly', () => {
-    const { rawClient1 } = rawClientFixture;
-    const expected = {
-      firstName: 'Khumpai',
-      gender: 'Male',
-      id: '9b67a82d-dac7-40c0-85aa-e5976339a6b6',
-      lastContactDate: '',
-      lastName: 'Family',
-      location: '',
-      middleName: '',
-      type: 'Client',
-    };
-    expect(extractClient(rawClient1)).toEqual(expected);
   });
 });
